deleting the test package

+++ GitProj01/src/main/java/com/nt/test/App.java
@@ -9,7 +9,7 @@ public class App {
     public static void main(String[] args) {
     	
     	Arithmetic ar = new Arithmetic();
-    	int result=ar.sum(10, 20);
+    	long result=ar.sum(10, 20);
         System.out.println(result);
     }
 }
